reorganize code a bit
