+++ src/config/network_config.py
@@ -1,0 +1,5 @@
+
+# Maximum number of attempts to connect
+MAX_ATTEMPT_COUNT = 3
+
+DEFAULT_TIMEOUT = 30 # s
