+++ src/util/simplify_read_id.py
